pre final fe

+++ src/pages/Admin/ViewRequest/ViewCourse.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import './ViewCourse.css';
 import authApi from '../../../api/authApi';
-import { Button, Space, Table } from 'antd';
+import { Button, Space, Table, Input } from 'antd';
 import moment from 'moment';
 import jwt_decode from 'jwt-decode';
 import Sidebar from '../../../components/Sidebar/Sidebar';
@@ -13,6 +13,7 @@ const ViewCourse = () => {
   const [lessons, setLessons] = useState([]);
   const [quizzes, setQuizzes] = useState([]);
   const navigate = useNavigate();
+  const [searchText, setSearchText] = useState('');
 
   const [user, setUser] = useState('');
 
@@ -23,6 +24,7 @@ const ViewCourse = () => {
       setUser(deCoded.sub);
     }
   }, []);
+
   useEffect(() => {
     authApi
       .getCourseById(courseID)
@@ -53,10 +55,15 @@ const ViewCourse = () => {
   }, []);
 
   const handleAddLesson = () => {
+    if (!localStorage.getItem('user-access-token')) return (window.location.href = '/signin');
+
     navigate(`/add-lesson/${courseID}`);
   };
 
   const handleDeleteLesson = (lessonID) => {
+    if (!localStorage.getItem('user-access-token')) return (window.location.href = '/signin');
+    console.log(lessonID);
+
     if (window.confirm('Do you want to delete this lesson?')) {
       authApi
         .deleteLesson(lessonID)
@@ -69,17 +76,26 @@ const ViewCourse = () => {
   };
 
   const handleEditLesson = (id) => {
+    if (!localStorage.getItem('user-access-token')) return (window.location.href = '/signin');
+
     navigate(`/edit-lesson/${id}`);
   };
+
   const handleViewLesson = (id) => {
+    if (!localStorage.getItem('user-access-token')) return (window.location.href = '/signin');
+
     navigate(`/viewLesson/${id}`);
   };
 
   const handleEditQuiz = (id) => {
+    if (!localStorage.getItem('user-access-token')) return (window.location.href = '/signin');
+
     navigate(`/edit-quiz/${id}`);
   };
 
   const handleDeleteQuiz = (quizID) => {
+    if (!localStorage.getItem('user-access-token')) return (window.location.href = '/signin');
+
     const params = {
       username: user,
       quizID,
@@ -94,10 +110,16 @@ const ViewCourse = () => {
         .catch((error) => {});
     }
   };
+
   const handleViewQuiz = (id) => {
+    if (!localStorage.getItem('user-access-token')) return (window.location.href = '/signin');
+
     navigate(`/view-quiz/${id}`);
   };
+
   const handleAddQuiz = (id) => {
+    if (!localStorage.getItem('user-access-token')) return (window.location.href = '/signin');
+
     navigate(`/add-quiz/${id}`);
   };
 
@@ -142,6 +164,12 @@ const ViewCourse = () => {
     },
   ];
 
+  const handleSearch = (value) => {
+    setSearchText(value);
+  };
+
+  const filteredLessons = lessons.filter((lesson) => lesson.name.toLowerCase().includes(searchText.toLowerCase()));
+
   return (
     <div style={{ display: 'flex' }}>
       <Sidebar />
@@ -175,11 +203,18 @@ const ViewCourse = () => {
           )}
 
           <h2>List Lesson</h2>
-          <button onClick={handleAddLesson}>Add Lesson</button>
-
+          <Button onClick={handleAddLesson} style={{ width: '100px' }}>
+            Add Lesson
+          </Button>
+          <Input.Search
+            placeholder="Search by lesson name"
+            allowClear
+            onSearch={handleSearch}
+            style={{ width: 200, marginBottom: 16 }}
+          />
           <Table
             columns={columns}
-            dataSource={lessons}
+            dataSource={filteredLessons}
             rowKey={(record) => record.id}
             expandable={{
               expandedRowRender: (record) => {
